refactor: 주문하기 첫 화면을 메뉴 리스트 뷰로 변경 및 우상단 네비 버튼을 장바구니로 변경
